Fixed delete user, userlist works.

+++ music-app/src/main/java/com/example/musicapp/service/UserService.java
@@ -8,7 +8,7 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 public interface UserService extends UserDetailsService {
     public Iterable<User> listUsers();
     public String createUser(User newUser);
-    public HttpStatus deleteById(Long userId);
+    public HttpStatus deleteById(Integer userId);
     public String login(User user);
     public User getUser(String username); //added for UserProfileImpl line 20 getUser
     public User addSong(String username, int songId);
